@@ -2,28 +2,28 @@ part of amap_core;
 
 class InputTip {
   /// 区域编码
-  final String adcode;
+  final String? adcode;
 
   /// 地址
-  final String address;
+  final String? address;
 
   /// 所属区域
-  final String district;
+  final String? district;
 
   /// 名称
-  final String name;
+  final String? name;
 
   /// 类型编码
-  final String typecode;
+  final String? typecode;
 
   /// POI全局唯一ID
-  final String uid;
+  final String? uid;
 
   /// 纬度（垂直方向）
-  final double latitude;
+  final double? latitude;
 
   /// 经度（水平方向）
-  final double longitude;
+  final double? longitude;
 
   InputTip(
     this.adcode,
@@ -42,8 +42,8 @@ class InputTip {
 }
 
 InputTip _$InputTipFromJson(Map<dynamic, dynamic> json) {
-  return InputTip(json['uid'] as String, json['name'] as String, json['typecode'] as String, json['address'] as String,
-      json['district'] as String, json['adcode'] as String, (json['latitude'] as num)?.toDouble(), (json['longitude'] as num)?.toDouble());
+  return InputTip(json['uid'] as String?, json['name'] as String?, json['typecode'] as String?, json['address'] as String?,
+      json['district'] as String?, json['adcode'] as String?, (json['latitude'] as num?)?.toDouble(), (json['longitude'] as num?)?.toDouble());
 }
 
 Map<String, dynamic> _$InputTipToJson(InputTip instance) => <String, dynamic>{
